refactor: add teamwork logic to separate package
