+++ teamwork/calculateStrategies.go
@@ -1,4 +1,4 @@
-package main
+package teamwork
 
 import "strconv"
