+++ teamwork/logTime.go
@@ -1,4 +1,4 @@
-package main
+package teamwork
 
 import (
 	"bufio"
@@ -37,14 +37,14 @@ func prepareData(projectMode bool) (*LogTimeMetaData, error) {
 	}
 
 	fmt.Printf("Enter %s: \n", hoursReference)
-	_, taskError := fmt.Scan(&taskId)
-	if taskError != nil {
-		return nil, taskError
+	_, err := fmt.Scan(&taskId)
+	if err != nil {
+		return nil, err
 	}
 
 	fmt.Println("Enter hours and minutes (separate with space): ")
 	_, hoursMinutesError := fmt.Scan(&hours, &minutes)
-	if taskError != nil {
+	if hoursMinutesError != nil {
 		return nil, hoursMinutesError
 	}
 
@@ -75,7 +75,7 @@ func prepareDataForRequest(workDays []time.Time, logMetadata *LogTimeMetaData, c
 	return &timeLogs
 }
 
-func logHours(startDate, endDate, startTime string, projectMode, includeCroHolidays, nonBillable bool, configuration config.Config) {
+func LogHours(startDate, endDate, startTime string, projectMode, includeCroHolidays, nonBillable bool, configuration config.Config) {
 	logMetadata, err := prepareData(projectMode)
 	if err != nil {
 		fmt.Printf("Error: %s", err)
